Update to pulling and uploading to GUI
The GUI will now load first names into the user section. The next thing to do is to be able to add plant sensors to said users. After that the sensor data can be compared to ideal values. After this is done, a second sensor should be setup to work on encapsulation of different user's data. After some UI work is done to look at the comparisons it will be time to organize the file structure. A batch file may be fun to make, but not useful. If things are running smoothly it's either time to consider an alert system or the app UI using React Native. The file cleanup, after the Proof of Concept is complete should be a high priority.